backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Base de datos y ORM espacial
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.14.7
alembic==1.13.1

# Autenticación JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validación y schemas
pydantic==2.7.1
pydantic-settings==2.2.1

# Motor hidráulico (EPANET)
wntr==1.2.0
numpy==1.26.4
pandas==2.2.2
networkx==3.3

# Importación / exportación de datos espaciales
fiona==1.9.6
shapely==2.0.4
pyproj==3.6.1

# Exportación de reportes
openpyxl==3.1.2
weasyprint==62.1

# Utilidades
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
